Detach CtClass instances from the ClassPool as they are inspected by the Javassist transform.
It is not necessary to keep them in the ClassPool and detaching them can reduce memory consumption and avoid out of memory errors.
Note that this only affects users with AGP older than 4.2.0 since the HiltTransformTestClassesTask was replaced by the ASM pipeline.

Fixes: https://github.com/google/dagger/issues/2567
RELNOTES=Reduce the possibility of OOMs in the local unit test transform task.
PiperOrigin-RevId: 369978156